+++ tests/integration/zipkin_integration_test.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import json
 
+import mock
 import pytest
 
 from py_zipkin import Encoding
@@ -8,8 +9,9 @@ from py_zipkin import Kind
 from py_zipkin import zipkin
 from py_zipkin.logging_helper import LOGGING_END_KEY
 from py_zipkin.storage import get_default_tracer
+from py_zipkin.zipkin import log
 from py_zipkin.zipkin import ZipkinAttrs
-
+from tests.test_helpers import MockTracer
 
 USECS = 1000000
 
@@ -249,6 +251,39 @@ def test_service_span(encoding):
         assert 'duration' not in span
     elif encoding == Encoding.V2_JSON:
         assert span['shared'] is True
+
+
+@mock.patch.object(log, 'error', autospec=True)
+def test_service_exits_on_erroneous_span(log_mock):
+    """Tests that for invalid zipkin_attrs exceptions are not thrown
+       Services may not be handling them. Instead log an error
+    """
+    mock_transport_handler, mock_logs = mock_logger()
+    tracer = MockTracer()
+    try:
+        zipkin_attrs = ZipkinAttrs(
+            trace_id='0',
+            span_id='1, 1',  # invalid span id
+            parent_span_id='2',
+            flags='0',
+            is_sampled=True,
+        )
+        with tracer.zipkin_span(
+            service_name='test_service_name',
+            span_name='service_span',
+            zipkin_attrs=zipkin_attrs,
+            transport_handler=mock_transport_handler,
+            encoding=Encoding.V1_THRIFT,
+        ):
+            pass
+
+    except Exception:
+        pytest.fail('Exception not expected to be thrown!')
+
+    finally:
+        assert log_mock.call_count == 1
+        assert len(mock_logs) == 0
+        assert len(tracer.get_spans()) == 0
 
 
 def test_service_span_report_timestamp_override():